feat(isEmpty): add higher-order lettable version of isEmpty
